feat(step3): 리뷰 반영
