--- src/main/java/kitchenpos/order/infra/KitchenridersClient.java
@@ -1,8 +1,0 @@
-package kitchenpos.order.infra;
-
-import java.math.BigDecimal;
-import java.util.UUID;
-
-public interface KitchenridersClient {
-    void requestDelivery(UUID orderId, BigDecimal amount, String deliveryAddress);
-}
