+++ src/main/java/kitchenpos/orderDelivery/domain/JpaOrderDeliveryRepository.java
@@ -1,0 +1,8 @@
+package kitchenpos.orderDelivery.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface JpaOrderDeliveryRepository extends OrderDeliveryRepository, JpaRepository<OrderDelivery, UUID> {
+}
