+++ src/main/java/kitchenpos/orderDelivery/domain/OrderDelivery.java
@@ -1,0 +1,80 @@
+package kitchenpos.orderDelivery.domain;
+
+import kitchenpos.orderCommonDomain.OrderLineItem;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.UUID;
+
+@Table(name = "orders")
+@Entity
+public class OrderDelivery {
+    @Column(name = "id", columnDefinition = "binary(16)")
+    @Id
+    private UUID id;
+
+    @Column(name = "status", nullable = false)
+    @Enumerated(EnumType.STRING)
+    private OrderDeliveryStatus status;
+
+    @Column(name = "order_date_time", nullable = false)
+    private LocalDateTime orderDateTime;
+
+    @OneToMany(cascade = {CascadeType.PERSIST, CascadeType.MERGE})
+    @JoinColumn(
+        name = "order_id",
+        nullable = false,
+        columnDefinition = "binary(16)",
+        foreignKey = @ForeignKey(name = "fk_order_line_item_to_orders")
+    )
+    private List<OrderLineItem> orderLineItems;
+
+    @Column(name = "delivery_address")
+    private String deliveryAddress;
+
+
+    public OrderDelivery() {
+    }
+
+    public UUID getId() {
+        return id;
+    }
+
+    public void setId(final UUID id) {
+        this.id = id;
+    }
+
+    public OrderDeliveryStatus getStatus() {
+        return status;
+    }
+
+    public void setStatus(final OrderDeliveryStatus status) {
+        this.status = status;
+    }
+
+    public LocalDateTime getOrderDateTime() {
+        return orderDateTime;
+    }
+
+    public void setOrderDateTime(final LocalDateTime orderDateTime) {
+        this.orderDateTime = orderDateTime;
+    }
+
+    public List<OrderLineItem> getOrderLineItems() {
+        return orderLineItems;
+    }
+
+    public void setOrderLineItems(final List<OrderLineItem> orderLineItems) {
+        this.orderLineItems = orderLineItems;
+    }
+
+    public String getDeliveryAddress() {
+        return deliveryAddress;
+    }
+
+    public void setDeliveryAddress(final String deliveryAddress) {
+        this.deliveryAddress = deliveryAddress;
+    }
+
+}
