+++ src/main/java/kitchenpos/orderDelivery/domain/OrderDeliveryRepository.java
@@ -1,0 +1,14 @@
+package kitchenpos.orderDelivery.domain;
+
+
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+
+public interface OrderDeliveryRepository {
+    OrderDelivery save(OrderDelivery order);
+
+    Optional<OrderDelivery> findById(UUID id);
+
+    List<OrderDelivery> findAll();
+}
