+++ src/main/java/kitchenpos/orderDelivery/domain/OrderDeliveryStatus.java
@@ -1,0 +1,5 @@
+package kitchenpos.orderDelivery.domain;
+
+public enum OrderDeliveryStatus {
+    WAITING, ACCEPTED, SERVED, DELIVERING, DELIVERED, COMPLETED
+}
