+++ src/main/java/kitchenpos/orderDelivery/infra/DefaultKitchenridersClient.java
@@ -1,0 +1,13 @@
+package kitchenpos.orderDelivery.infra;
+
+import org.springframework.stereotype.Component;
+
+import java.math.BigDecimal;
+import java.util.UUID;
+
+@Component
+public class DefaultKitchenridersClient implements KitchenridersClient {
+    @Override
+    public void requestDelivery(final UUID orderId, final BigDecimal amount, final String deliveryAddress) {
+    }
+}
