+++ src/main/java/kitchenpos/orderDelivery/infra/KitchenridersClient.java
@@ -1,0 +1,8 @@
+package kitchenpos.orderDelivery.infra;
+
+import java.math.BigDecimal;
+import java.util.UUID;
+
+public interface KitchenridersClient {
+    void requestDelivery(UUID orderId, BigDecimal amount, String deliveryAddress);
+}
