+++ src/main/java/kitchenpos/orderEatIn/domain/JpaOrderEatInRepository.java
@@ -1,0 +1,8 @@
+package kitchenpos.orderEatIn.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface JpaOrderEatInRepository extends OrderEatInRepository, JpaRepository<OrderEatIn, UUID> {
+}
