+++ src/main/java/kitchenpos/orderEatIn/domain/JpaOrderTableRepository.java
@@ -1,0 +1,8 @@
+package kitchenpos.orderEatIn.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface JpaOrderTableRepository extends OrderTableRepository, JpaRepository<OrderTable, UUID> {
+}
