+++ src/main/java/kitchenpos/orderEatIn/domain/OrderEatInRepository.java
@@ -1,0 +1,15 @@
+package kitchenpos.orderEatIn.domain;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+
+public interface OrderEatInRepository {
+    OrderEatIn save(OrderEatIn order);
+
+    Optional<OrderEatIn> findById(UUID id);
+
+    List<OrderEatIn> findAll();
+
+    boolean existsByOrderTableAndStatusNot(OrderTable orderTable, OrderEatInStatus status);
+}
