+++ src/main/java/kitchenpos/orderEatIn/domain/OrderEatInStatus.java
@@ -1,0 +1,5 @@
+package kitchenpos.orderEatIn.domain;
+
+public enum OrderEatInStatus {
+    WAITING, ACCEPTED, SERVED, COMPLETED
+}
