--- src/main/java/kitchenpos/orderTable/domain/JpaOrderTableRepository.java
@@ -1,8 +1,0 @@
-package kitchenpos.orderTable.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface JpaOrderTableRepository extends OrderTableRepository, JpaRepository<OrderTable, UUID> {
-}
