--- src/main/java/kitchenpos/orderTable/domain/OrderTableRepository.java
@@ -1,14 +1,0 @@
-package kitchenpos.orderTable.domain;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-
-public interface OrderTableRepository {
-    OrderTable save(OrderTable orderTable);
-
-    Optional<OrderTable> findById(UUID id);
-
-    List<OrderTable> findAll();
-}
-
