+++ src/main/java/kitchenpos/orderTakeOut/domain/JpaOrderTakeOutRepository.java
@@ -1,0 +1,8 @@
+package kitchenpos.orderTakeOut.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface JpaOrderTakeOutRepository extends OrderTakeOutRepository, JpaRepository<OrderTakeOut, UUID> {
+}
