+++ src/main/java/kitchenpos/orderTakeOut/domain/OrderTakeOutStatus.java
@@ -1,0 +1,5 @@
+package kitchenpos.orderTakeOut.domain;
+
+public enum OrderTakeOutStatus {
+    WAITING, ACCEPTED, SERVED, COMPLETED
+}
