+++ src/test/java/kitchenpos/application/InMemoryOrderEatInRepository.java
@@ -1,0 +1,35 @@
+package kitchenpos.application;
+
+import kitchenpos.orderEatIn.domain.OrderEatIn;
+import kitchenpos.orderEatIn.domain.OrderEatInRepository;
+import kitchenpos.orderEatIn.domain.OrderEatInStatus;
+import kitchenpos.orderEatIn.domain.OrderTable;
+
+import java.util.*;
+
+public class InMemoryOrderEatInRepository implements OrderEatInRepository {
+    private final Map<UUID, OrderEatIn> orders = new HashMap<>();
+
+    @Override
+    public OrderEatIn save(final OrderEatIn order) {
+        orders.put(order.getId(), order);
+        return order;
+    }
+
+    @Override
+    public Optional<OrderEatIn> findById(final UUID id) {
+        return Optional.ofNullable(orders.get(id));
+    }
+
+    @Override
+    public List<OrderEatIn> findAll() {
+        return new ArrayList<>(orders.values());
+    }
+
+    @Override
+    public boolean existsByOrderTableAndStatusNot(final OrderTable orderTable, final OrderEatInStatus status) {
+        return orders.values()
+            .stream()
+            .anyMatch(order -> order.getOrderTable().equals(orderTable) && order.getStatus() != status);
+    }
+}
