+++ src/test/java/kitchenpos/application/OrderDeliveryServiceTest.java
@@ -1,0 +1,246 @@
+package kitchenpos.application;
+
+import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.orderCommonDomain.OrderLineItem;
+import kitchenpos.orderDelivery.application.OrderDeliveryService;
+import kitchenpos.orderDelivery.domain.OrderDelivery;
+import kitchenpos.orderDelivery.domain.OrderDeliveryRepository;
+import kitchenpos.orderDelivery.domain.OrderDeliveryStatus;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.*;
+
+import java.math.BigDecimal;
+import java.util.*;
+
+import static kitchenpos.Fixtures.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.junit.jupiter.api.Assertions.assertAll;
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+
+class OrderDeliveryServiceTest {
+    private OrderDeliveryRepository orderRepository;
+    private MenuRepository menuRepository;
+    private FakeKitchenridersClient kitchenridersClient;
+    private OrderDeliveryService orderService;
+
+    @BeforeEach
+    void setUp() {
+        orderRepository = new InMemoryOrderDeliveryRepository();
+        menuRepository = new InMemoryMenuRepository();
+        kitchenridersClient = new FakeKitchenridersClient();
+        orderService = new OrderDeliveryService(orderRepository, menuRepository, kitchenridersClient);
+    }
+
+    @DisplayName("1개 이상의 등록된 메뉴로 배달 주문을 등록할 수 있다.")
+    @Test
+    void createDeliveryOrder() {
+        final UUID menuId = menuRepository.save(menu(19_000L, true, menuProduct())).getId();
+        final OrderDelivery expected = createOrderRequest(
+            "서울시 송파구 위례성대로 2", createOrderLineItemRequest(menuId, 19_000L, 3L)
+        );
+        final OrderDelivery actual = orderService.create(expected);
+        assertThat(actual).isNotNull();
+        assertAll(
+            () -> assertThat(actual.getId()).isNotNull(),
+            () -> assertThat(actual.getOrderDateTime()).isNotNull(),
+            () -> assertThat(actual.getOrderLineItems()).hasSize(1),
+            () -> assertThat(actual.getDeliveryAddress()).isEqualTo(expected.getDeliveryAddress())
+        );
+    }
+
+
+
+    @DisplayName("메뉴가 없으면 등록할 수 없다.")
+    @MethodSource("orderLineItems")
+    @ParameterizedTest
+    void create(final List<OrderLineItem> orderLineItems) {
+        final OrderDelivery expected = createOrderRequest( orderLineItems);
+        assertThatThrownBy(() -> orderService.create(expected))
+            .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    private static List<Arguments> orderLineItems() {
+        return Arrays.asList(
+            null,
+            Arguments.of(Collections.emptyList()),
+            Arguments.of(Arrays.asList(createOrderLineItemRequest(INVALID_ID, 19_000L, 3L)))
+        );
+    }
+
+    @DisplayName("매장 주문을 제외한 주문의 경우 주문 항목의 수량은 0 이상이어야 한다.")
+    @ValueSource(longs = -1L)
+    @ParameterizedTest
+    void createWithoutEatInOrder(final long quantity) {
+        final UUID menuId = menuRepository.save(menu(19_000L, true, menuProduct())).getId();
+        final OrderDelivery expected = createOrderRequest(createOrderLineItemRequest(menuId, 19_000L, quantity)
+        );
+        assertThatThrownBy(() -> orderService.create(expected))
+            .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @DisplayName("배달 주소가 올바르지 않으면 배달 주문을 등록할 수 없다.")
+    @NullAndEmptySource
+    @ParameterizedTest
+    void create(final String deliveryAddress) {
+        final UUID menuId = menuRepository.save(menu(19_000L, true, menuProduct())).getId();
+        final OrderDelivery expected = createOrderRequest(deliveryAddress, createOrderLineItemRequest(menuId, 19_000L, 3L)
+        );
+        assertThatThrownBy(() -> orderService.create(expected))
+            .isInstanceOf(IllegalArgumentException.class);
+    }
+
+
+
+    @DisplayName("숨겨진 메뉴는 주문할 수 없다.")
+    @Test
+    void createNotDisplayedMenuOrder() {
+        final UUID menuId = menuRepository.save(menu(19_000L, false, menuProduct())).getId();
+        final OrderDelivery expected = createOrderRequest( createOrderLineItemRequest(menuId, 19_000L, 3L));
+        assertThatThrownBy(() -> orderService.create(expected))
+            .isInstanceOf(IllegalStateException.class);
+    }
+
+    @DisplayName("주문한 메뉴의 가격은 실제 메뉴 가격과 일치해야 한다.")
+    @Test
+    void createNotMatchedMenuPriceOrder() {
+        final UUID menuId = menuRepository.save(menu(19_000L, true, menuProduct())).getId();
+        final OrderDelivery expected = createOrderRequest( createOrderLineItemRequest(menuId, 16_000L, 3L));
+        assertThatThrownBy(() -> orderService.create(expected))
+            .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @DisplayName("주문을 접수한다.")
+    @Test
+    void accept() {
+        final UUID orderId = orderRepository.save(orderDelivery(OrderDeliveryStatus.WAITING)).getId();
+        final OrderDelivery actual = orderService.accept(orderId);
+        assertThat(actual.getStatus()).isEqualTo(OrderDeliveryStatus.ACCEPTED);
+    }
+
+    @DisplayName("접수 대기 중인 주문만 접수할 수 있다.")
+    @EnumSource(value = OrderDeliveryStatus.class, names = "WAITING", mode = EnumSource.Mode.EXCLUDE)
+    @ParameterizedTest
+    void accept(final OrderDeliveryStatus status) {
+        final UUID orderId = orderRepository.save(orderDelivery(status)).getId();
+        assertThatThrownBy(() -> orderService.accept(orderId))
+            .isInstanceOf(IllegalStateException.class);
+    }
+
+    @DisplayName("배달 주문을 접수되면 배달 대행사를 호출한다.")
+    @Test
+    void acceptDeliveryOrder() {
+        final UUID orderId = orderRepository.save(orderDelivery(OrderDeliveryStatus.WAITING, "서울시 송파구 위례성대로 2")).getId();
+        final OrderDelivery actual = orderService.accept(orderId);
+        assertAll(
+            () -> assertThat(actual.getStatus()).isEqualTo(OrderDeliveryStatus.ACCEPTED),
+            () -> assertThat(kitchenridersClient.getOrderId()).isEqualTo(orderId),
+            () -> assertThat(kitchenridersClient.getDeliveryAddress()).isEqualTo("서울시 송파구 위례성대로 2")
+        );
+    }
+
+    @DisplayName("주문을 서빙한다.")
+    @Test
+    void serve() {
+        final UUID orderId = orderRepository.save(orderDelivery(OrderDeliveryStatus.ACCEPTED)).getId();
+        final OrderDelivery actual = orderService.serve(orderId);
+        assertThat(actual.getStatus()).isEqualTo(OrderDeliveryStatus.SERVED);
+    }
+
+    @DisplayName("접수된 주문만 서빙할 수 있다.")
+    @EnumSource(value = OrderDeliveryStatus.class, names = "ACCEPTED", mode = EnumSource.Mode.EXCLUDE)
+    @ParameterizedTest
+    void serve(final OrderDeliveryStatus status) {
+        final UUID orderId = orderRepository.save(orderDelivery(status)).getId();
+        assertThatThrownBy(() -> orderService.serve(orderId))
+            .isInstanceOf(IllegalStateException.class);
+    }
+
+    @DisplayName("주문을 배달한다.")
+    @Test
+    void startDelivery() {
+        final UUID orderId = orderRepository.save(orderDelivery(OrderDeliveryStatus.SERVED, "서울시 송파구 위례성대로 2")).getId();
+        final OrderDelivery actual = orderService.startDelivery(orderId);
+        assertThat(actual.getStatus()).isEqualTo(OrderDeliveryStatus.DELIVERING);
+    }
+
+
+    @DisplayName("서빙된 주문만 배달할 수 있다.")
+    @EnumSource(value = OrderDeliveryStatus.class, names = "SERVED", mode = EnumSource.Mode.EXCLUDE)
+    @ParameterizedTest
+    void startDelivery(final OrderDeliveryStatus status) {
+        final UUID orderId = orderRepository.save(orderDelivery(status, "서울시 송파구 위례성대로 2")).getId();
+        assertThatThrownBy(() -> orderService.startDelivery(orderId))
+            .isInstanceOf(IllegalStateException.class);
+    }
+
+    @DisplayName("주문을 배달 완료한다.")
+    @Test
+    void completeDelivery() {
+        final UUID orderId = orderRepository.save(orderDelivery(OrderDeliveryStatus.DELIVERING, "서울시 송파구 위례성대로 2")).getId();
+        final OrderDelivery actual = orderService.completeDelivery(orderId);
+        assertThat(actual.getStatus()).isEqualTo(OrderDeliveryStatus.DELIVERED);
+    }
+
+    @DisplayName("배달 중인 주문만 배달 완료할 수 있다.")
+    @EnumSource(value = OrderDeliveryStatus.class, names = "DELIVERING", mode = EnumSource.Mode.EXCLUDE)
+    @ParameterizedTest
+    void completeDelivery(final OrderDeliveryStatus status) {
+        final UUID orderId = orderRepository.save(orderDelivery(status, "서울시 송파구 위례성대로 2")).getId();
+        assertThatThrownBy(() -> orderService.completeDelivery(orderId))
+            .isInstanceOf(IllegalStateException.class);
+    }
+
+    @DisplayName("주문을 완료한다.")
+    @Test
+    void complete() {
+        final OrderDelivery expected = orderRepository.save(orderDelivery(OrderDeliveryStatus.DELIVERED, "서울시 송파구 위례성대로 2"));
+        final OrderDelivery actual = orderService.complete(expected.getId());
+        assertThat(actual.getStatus()).isEqualTo(OrderDeliveryStatus.COMPLETED);
+    }
+
+    @DisplayName("배달 주문의 경우 배달 완료된 주문만 완료할 수 있다.")
+    @EnumSource(value = OrderDeliveryStatus.class, names = "DELIVERED", mode = EnumSource.Mode.EXCLUDE)
+    @ParameterizedTest
+    void completeDeliveryOrder(final OrderDeliveryStatus status) {
+        final UUID orderId = orderRepository.save(orderDelivery(status, "서울시 송파구 위례성대로 2")).getId();
+        assertThatThrownBy(() -> orderService.complete(orderId))
+            .isInstanceOf(IllegalStateException.class);
+    }
+
+
+
+    private OrderDelivery createOrderRequest(
+        final String deliveryAddress,
+        final OrderLineItem... orderLineItems
+    ) {
+        final OrderDelivery order = new OrderDelivery();
+        order.setDeliveryAddress(deliveryAddress);
+        order.setOrderLineItems(Arrays.asList(orderLineItems));
+        return order;
+    }
+
+    private OrderDelivery createOrderRequest( final OrderLineItem... orderLineItems) {
+        return createOrderRequest( Arrays.asList(orderLineItems));
+    }
+
+    private OrderDelivery createOrderRequest(final List<OrderLineItem> orderLineItems) {
+        final OrderDelivery order = new OrderDelivery();
+        order.setOrderLineItems(orderLineItems);
+        return order;
+    }
+
+
+
+    private static OrderLineItem createOrderLineItemRequest(final UUID menuId, final long price, final long quantity) {
+        final OrderLineItem orderLineItem = new OrderLineItem();
+        orderLineItem.setSeq(new Random().nextLong());
+        orderLineItem.setMenuId(menuId);
+        orderLineItem.setPrice(BigDecimal.valueOf(price));
+        orderLineItem.setQuantity(quantity);
+        return orderLineItem;
+    }
+}
