+++ src/test/java/kitchenpos/application/OrderTakeOutServiceTest.java
@@ -1,0 +1,154 @@
+package kitchenpos.application;
+
+import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.orderCommonDomain.OrderLineItem;
+import kitchenpos.orderTakeOut.application.OrderTakeOutService;
+import kitchenpos.orderTakeOut.domain.OrderTakeOut;
+import kitchenpos.orderTakeOut.domain.OrderTakeOutRepository;
+import kitchenpos.orderTakeOut.domain.OrderTakeOutStatus;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.*;
+
+import java.math.BigDecimal;
+import java.util.*;
+
+import static kitchenpos.Fixtures.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.junit.jupiter.api.Assertions.assertAll;
+
+class OrderTakeOutServiceTest {
+    private OrderTakeOutRepository orderRepository;
+    private MenuRepository menuRepository;
+    private OrderTakeOutService orderService;
+
+    @BeforeEach
+    void setUp() {
+        orderRepository = new InMemoryOrderTakeOutRepository();
+        menuRepository = new InMemoryMenuRepository();
+        orderService = new OrderTakeOutService(orderRepository, menuRepository);
+    }
+
+
+    @DisplayName("1개 이상의 등록된 메뉴로 포장 주문을 등록할 수 있다.")
+    @Test
+    void createTakeoutOrder() {
+        final UUID menuId = menuRepository.save(menu(19_000L, true, menuProduct())).getId();
+        final OrderTakeOut expected = createOrderRequest(createOrderLineItemRequest(menuId, 19_000L, 3L));
+        final OrderTakeOut actual = orderService.create(expected);
+        assertThat(actual).isNotNull();
+        assertAll(
+            () -> assertThat(actual.getId()).isNotNull(),
+            () -> assertThat(actual.getStatus()).isEqualTo(OrderTakeOutStatus.WAITING),
+            () -> assertThat(actual.getOrderDateTime()).isNotNull(),
+            () -> assertThat(actual.getOrderLineItems()).hasSize(1)
+        );
+    }
+
+    @DisplayName("메뉴가 없으면 등록할 수 없다.")
+    @MethodSource("orderLineItems")
+    @ParameterizedTest
+    void create(final List<OrderLineItem> orderLineItems) {
+        final OrderTakeOut expected = createOrderRequest(orderLineItems);
+        assertThatThrownBy(() -> orderService.create(expected))
+            .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    private static List<Arguments> orderLineItems() {
+        return Arrays.asList(
+            null,
+            Arguments.of(Collections.emptyList()),
+            Arguments.of(Arrays.asList(createOrderLineItemRequest(INVALID_ID, 19_000L, 3L)))
+        );
+    }
+
+
+
+    @DisplayName("매장 주문을 제외한 주문의 경우 주문 항목의 수량은 0 이상이어야 한다.")
+    @ValueSource(longs = -1L)
+    @ParameterizedTest
+    void createWithoutEatInOrder(final long quantity) {
+        final UUID menuId = menuRepository.save(menu(19_000L, true, menuProduct())).getId();
+        final OrderTakeOut expected = createOrderRequest(
+             createOrderLineItemRequest(menuId, 19_000L, quantity)
+        );
+        assertThatThrownBy(() -> orderService.create(expected))
+            .isInstanceOf(IllegalArgumentException.class);
+    }
+
+
+
+    @DisplayName("숨겨진 메뉴는 주문할 수 없다.")
+    @Test
+    void createNotDisplayedMenuOrder() {
+        final UUID menuId = menuRepository.save(menu(19_000L, false, menuProduct())).getId();
+        final OrderTakeOut expected = createOrderRequest( createOrderLineItemRequest(menuId, 19_000L, 3L));
+        assertThatThrownBy(() -> orderService.create(expected))
+            .isInstanceOf(IllegalStateException.class);
+    }
+
+    @DisplayName("주문한 메뉴의 가격은 실제 메뉴 가격과 일치해야 한다.")
+    @Test
+    void createNotMatchedMenuPriceOrder() {
+        final UUID menuId = menuRepository.save(menu(19_000L, true, menuProduct())).getId();
+        final OrderTakeOut expected = createOrderRequest(createOrderLineItemRequest(menuId, 16_000L, 3L));
+        assertThatThrownBy(() -> orderService.create(expected))
+            .isInstanceOf(IllegalArgumentException.class);
+    }
+
+
+
+    @DisplayName("주문을 서빙한다.")
+    @Test
+    void serve() {
+        final UUID orderId = orderRepository.save(orderTakeOut(OrderTakeOutStatus.ACCEPTED)).getId();
+        final OrderTakeOut actual = orderService.serve(orderId);
+        assertThat(actual.getStatus()).isEqualTo(OrderTakeOutStatus.SERVED);
+    }
+
+    @DisplayName("접수된 주문만 서빙할 수 있다.")
+    @EnumSource(value = OrderTakeOutStatus.class, names = "ACCEPTED", mode = EnumSource.Mode.EXCLUDE)
+    @ParameterizedTest
+    void serve(final OrderTakeOutStatus status) {
+        final UUID orderId = orderRepository.save(orderTakeOut(status)).getId();
+        assertThatThrownBy(() -> orderService.serve(orderId))
+            .isInstanceOf(IllegalStateException.class);
+    }
+
+
+
+
+    @DisplayName("포장 및 매장 주문의 경우 서빙된 주문만 완료할 수 있다.")
+    @EnumSource(value = OrderTakeOutStatus.class, names = "SERVED", mode = EnumSource.Mode.EXCLUDE)
+    @ParameterizedTest
+    void completeTakeoutAndEatInOrder(final OrderTakeOutStatus status) {
+        final UUID orderId = orderRepository.save(orderTakeOut(status)).getId();
+        assertThatThrownBy(() -> orderService.complete(orderId))
+            .isInstanceOf(IllegalStateException.class);
+    }
+
+
+
+    private OrderTakeOut createOrderRequest( final OrderLineItem... orderLineItems) {
+        return createOrderRequest( Arrays.asList(orderLineItems));
+    }
+
+    private OrderTakeOut createOrderRequest( final List<OrderLineItem> orderLineItems) {
+        final OrderTakeOut order = new OrderTakeOut();
+        order.setOrderLineItems(orderLineItems);
+        return order;
+    }
+
+
+    private static OrderLineItem createOrderLineItemRequest(final UUID menuId, final long price, final long quantity) {
+        final OrderLineItem orderLineItem = new OrderLineItem();
+        orderLineItem.setSeq(new Random().nextLong());
+        orderLineItem.setMenuId(menuId);
+        orderLineItem.setPrice(BigDecimal.valueOf(price));
+        orderLineItem.setQuantity(quantity);
+        return orderLineItem;
+    }
+}
